Change to better branch names
The idea of having branch names of the form "issue-#" is not a good one. We changed the branch names to "prefix-#-short-title-description", where prefix indicates a group, like doc, bug, feature, or idea.

We also removed the "-rc" suffix from tags because everything in the main branch is considered deployable.